Check for file errors and fix path creation bug
Check for null before closing files to avoid crashes

Also there was a bug in the path creation function where strlen was
being called on a non-null terminated string which caused the length
determination to fail and for garbage to be added to the end of the
path. This caused the last char in the todo data file to be crap.