Continue working on add command
This commit fixes some issues with reading in data relating to whether
or not to use defined state in a new todo. When we unfortunatly use
scanf to get the priority the input buffer was left polluted with a \n
char which would then cause the next call to input_to_bool to instantly
return since it assumed that the user was choosing the default option.

This was fixed by adding a function which flushes stdin after calling
validate_scan_result

Signed-off-by: Chris Frank <chris@cfrank.org>